Moved app.js to app folder
