+++ app/app.js
@@ -7,15 +7,15 @@ var logger = require('morgan');
 var cookieParser = require('cookie-parser');
 var bodyParser = require('body-parser');
 
-var index = require('./routes/index');
-var users = require('./routes/users');
+var index = require('./../routes/index');
+var users = require('./../routes/users');
 
 var app = express();
-var db = require('./model/db');
-var User = require('./model/users');
+var db = require('./../model/db');
+var User = require('./../model/users');
 
 // view engine setup
-app.set('views', path.join(__dirname, 'views'));
+app.set('views', path.join(__dirname + '/..', 'views'));
 app.set('view engine', 'jade');
 
 // uncomment after placing your favicon in /public
