+++ bin/debug.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 
-export DEBUG=my-server:*
+export DEBUG=app:*
 node "${DIR}"/www.js
